Move structures.rs to api_structures.rs to state where they are used
